requirements.txt
msal>=1.20.0
requests>=2.28.0
tqdm>=4.64.0
python-dotenv>=0.19.0
